LayoutTests/platform/win/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/input-textselection-01-expected.txt


PASS The selection interface members
PASS Selection attributes should apply to type text
PASS Selection attributes should apply to type search
PASS Selection attributes should apply to type url
PASS Selection attributes should apply to type tel
PASS Selection attributes should apply to type email
PASS Selection attributes should apply to type password
FAIL Selection attributes should not apply to type hidden assert_throws: Should throw with type hidden function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type datetime assert_equals: Input type unsupported expected "datetime" but got "text"
FAIL Selection attributes should not apply to type date assert_equals: Input type unsupported expected "date" but got "text"
FAIL Selection attributes should not apply to type month assert_equals: Input type unsupported expected "month" but got "text"
FAIL Selection attributes should not apply to type week assert_equals: Input type unsupported expected "week" but got "text"
FAIL Selection attributes should not apply to type datetime-local assert_equals: Input type unsupported expected "datetime-local" but got "text"
FAIL Selection attributes should not apply to type number assert_throws: Should throw with type number function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type range assert_throws: Should throw with type range function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type color assert_equals: Input type unsupported expected "color" but got "text"
FAIL Selection attributes should not apply to type checkbox assert_throws: Should throw with type checkbox function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type radio assert_throws: Should throw with type radio function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type button assert_throws: Should throw with type button function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type file assert_throws: Should throw with type file function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type submit assert_throws: Should throw with type submit function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type image assert_throws: Should throw with type image function "function () { input.select(); }" did not throw
FAIL Selection attributes should not apply to type reset assert_throws: Should throw with type reset function "function () { input.select(); }" did not throw
